refactor(#414): extract AgentAvatar, UserAvatar, MarkdownContent from ChatConsole
- Create features/chat/components/Avatars.tsx (AgentAvatar + UserAvatar)
- Create features/chat/components/MarkdownContent.tsx (stripThinkBlocks + MarkdownContent)
- Remove ~156 lines from ChatConsole.tsx

Part 1 of ChatConsole God Component split. More sub-components to follow.
