+++ apps/desktop/src/renderer/features/chat/components/Avatars.tsx
@@ -1,0 +1,21 @@
+export function AgentAvatar() {
+  return (
+    <div
+      className="w-8 h-8 rounded-full flex items-center justify-center shrink-0 text-xs font-bold text-white mt-0.5"
+      style={{ background: 'var(--accent)' }}
+    >
+      A
+    </div>
+  );
+}
+
+export function UserAvatar() {
+  return (
+    <div
+      className="w-8 h-8 rounded-full flex items-center justify-center shrink-0 text-xs font-bold text-white mt-0.5"
+      style={{ background: 'var(--avatar-dark)' }}
+    >
+      U
+    </div>
+  );
+}
